Update OpenAIVisionParserPromptSchema
# Update OpenAIVisionParserPromptSchema


Summary: The schema was added a while ago to allow us to model the multimodal schema in the config. Now that we have the parser, some updates are needed (e.g. fix the settings supported since they differ a bit from the other chat models).

Test Plan: Input and settings render correctly
![Screenshot 2024-03-07 at 5 07 20 PM](https://github.com/lastmile-ai/aiconfig/assets/5060851/475abb6b-6fe1-40fa-b7ab-e71648d0b92b)