Implement OpenAIVisionParser
# Implement OpenAIVisionParser

Summary: Using the default openai parser as the basis (and extending it), implement a parser for text-and-image openai models (e.g. gpt-4-v). The key differences are:
- vision models don't support function calling or tools, so exclude from the completion params
- serialize/deserialize need to be a bit different to construct the user messages' content with text and image urls

One thing to note is the default max tokens for GPT-4V is quite small. Bumping it to 100 in #1419

Another thing I found while testing is that the model gets confused sometimes when you refer to the attached image as an image. For example, I had a prompt as "What is this image of" and it responded with "I can't provide information on any images until you upload one". Then I reran a handful of times and it finally gave expected result. Changing the prompt to "What is this?" seems to provide much better results.

Test Plan:

https://github.com/lastmile-ai/aiconfig/assets/5060851/47e0ae90-fab7-4d4f-98e9-163cd992aeed