[vscode] add session id to telemetry
Summary:

in an effort to understand distinct session users, this diff adds a unique session identifier to telemetry.

Unfortunately, it looks like the "default" sessionID that comes with datadog browser telemetry is not being sent, similar to what we noticed with [gradio](https://github.com/lastmile-ai/gradio-notebook/pull/184). Hence the need to define this explicitly

Note: This does not collect new metrics. This will allow for the LastMile team to understand how many users are using our vscode extension.

Test Plan: