[1/n] Move Global Parameters into ConfigMetadataContainer
# [1/n] Move Global Parameters into ConfigMetadataContainer

This PR stack will set up a basic UI for managing global model settings. Up until now, we've had no way to set model settings (such as system prompt) in the global config metadata from within the editor UI. We've had user feedback about this pain-point, and this is also going to be a crucial integration for the eval UI we're working on (since otherwise you would need to manually add different system prompts to all prompts in the config being evaluated, which is a LOT more friction than current state of using a single system prompt variable in python).

In this PR:
- Refactor Global parameters component into a global ConfigMetadataContainer which itself contains the global parameters component
- Subsequent PRs will add global model settings to this container as well

## Testing:
- Just ensure global parameters work as normal

![Screenshot 2024-03-12 at 2 53 22 PM](https://github.com/lastmile-ai/aiconfig/assets/5060851/8e1f4dac-77c8-4388-b0a9-82b21e85690f)